Allow running a command with ssh session
When starting or joining an exec with ssh, allow passing
`-- <some user command> <some args>` to the ssh command. This will
execute the command on the exec and stream the logs back.

This changes the semantics of the -- separator from passing SSH
connection options to being the user command to run.

The connection options can now be set using the repeatable
`--connection-option=` flag. e.g.
`--connection-option=StrictHostKeyChecking=yes`