Add exec command
Add command that will run on the remote exec.
By default the ssh session will detach and leave the command running
using nohup. The Pid of the command will be stored in pid.nohup file. We
can use this later to stop the exec command if we need to.

Include the `--interactive`/`-i` flag to stay attached to the remote after
running the exec command.

Also:
- share some of the connection logic between SSH and Exec commands by
  introducing an interface that allows the two commands to differ.