Add exec command (#29)
* Add exec command

Add command that will run on the remote exec.
By default the ssh session will detach and leave the command running
using nohup. The Pid of the command will be stored in pid.nohup file. We
can use this later to stop the exec command if we need to.

Include the `--interactive` flag to stay attached to the remote after
running the exec command.

Also:
- share some of the connection logic between SSH and Exec commands by
  introducing an interface that allows the two commands to differ.

* Remove session id from exec

Remove the session-id from the exec command, so that exec command will
always create a new exec. Pass the command to the api to be stored along
with the details of the exec.

Also
- Wrap the user's command in `bash -c` allowing the user to input
  multiple commands in a single string. e.g. `cd somedir && ./serve.py`

* Ensure exec has node type flags